Clean up outputparser code
